client/requirements.txt
cryptography==45.0.4
pillow==11.3.0
requests==2.32.4